use local active schedule table
[#119511307]
- create local active schedule table for scaling engine
- merge historydb and activescheduledb to scalingenginedb
- save and remove active schedule when being notified of active schedule start/end
- return 404 when active schedule does not exist in local database
