+++ src/autoscaler/db/sqldb/scalingengine_sqldb.go
@@ -1,0 +1,197 @@
+package sqldb
+
+import (
+	"code.cloudfoundry.org/lager"
+	"database/sql"
+	_ "github.com/lib/pq"
+
+	"autoscaler/db"
+	"autoscaler/models"
+
+	"time"
+)
+
+type ScalingEngineSQLDB struct {
+	url    string
+	logger lager.Logger
+	sqldb  *sql.DB
+}
+
+func NewScalingEngineSQLDB(url string, logger lager.Logger) (*ScalingEngineSQLDB, error) {
+	sqldb, err := sql.Open(db.PostgresDriverName, url)
+	if err != nil {
+		logger.Error("open-scaling-engine-db", err, lager.Data{"url": url})
+		return nil, err
+	}
+
+	err = sqldb.Ping()
+	if err != nil {
+		sqldb.Close()
+		logger.Error("ping-scaling-engine-db", err, lager.Data{"url": url})
+		return nil, err
+	}
+
+	return &ScalingEngineSQLDB{
+		url:    url,
+		logger: logger,
+		sqldb:  sqldb,
+	}, nil
+}
+
+func (sdb *ScalingEngineSQLDB) Close() error {
+	err := sdb.sqldb.Close()
+	if err != nil {
+		sdb.logger.Error("close-scaling-engine-db", err, lager.Data{"url": sdb.url})
+		return err
+	}
+	return nil
+}
+
+func (sdb *ScalingEngineSQLDB) SaveScalingHistory(history *models.AppScalingHistory) error {
+	query := "INSERT INTO scalinghistory" +
+		"(appid, timestamp, scalingtype, status, oldinstances, newinstances, reason, message, error) " +
+		" VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9)"
+	_, err := sdb.sqldb.Exec(query, history.AppId, history.Timestamp, history.ScalingType, history.Status,
+		history.OldInstances, history.NewInstances, history.Reason, history.Message, history.Error)
+
+	if err != nil {
+		sdb.logger.Error("save-scaling-history", err, lager.Data{"query": query, "history": history})
+	}
+	return err
+}
+
+func (sdb *ScalingEngineSQLDB) RetrieveScalingHistories(appId string, start int64, end int64) ([]*models.AppScalingHistory, error) {
+	query := "SELECT timestamp, scalingtype, status, oldinstances, newinstances, reason, message, error FROM scalinghistory WHERE" +
+		" appid = $1 " +
+		" AND timestamp >= $2" +
+		" AND timestamp <= $3 ORDER BY timestamp"
+
+	if end < 0 {
+		end = time.Now().UnixNano()
+	}
+
+	histories := []*models.AppScalingHistory{}
+	rows, err := sdb.sqldb.Query(query, appId, start, end)
+	if err != nil {
+		sdb.logger.Error("retrieve-scaling-histories", err,
+			lager.Data{"query": query, "appid": appId, "start": start, "end": end})
+		return nil, err
+	}
+
+	defer rows.Close()
+
+	var timestamp int64
+	var scalingType, status, oldInstances, newInstances int
+	var reason, message, errorMsg string
+
+	for rows.Next() {
+		if err = rows.Scan(&timestamp, &scalingType, &status, &oldInstances, &newInstances, &reason, &message, &errorMsg); err != nil {
+			sdb.logger.Error("retrieve-scaling-history-scan", err)
+			return nil, err
+		}
+
+		history := models.AppScalingHistory{
+			AppId:        appId,
+			Timestamp:    timestamp,
+			ScalingType:  models.ScalingType(scalingType),
+			Status:       models.ScalingStatus(status),
+			OldInstances: oldInstances,
+			NewInstances: newInstances,
+			Reason:       reason,
+			Message:      message,
+			Error:        errorMsg,
+		}
+		histories = append(histories, &history)
+	}
+	return histories, nil
+}
+
+func (sdb *ScalingEngineSQLDB) CanScaleApp(appId string) (bool, error) {
+	query := "SELECT expireat FROM scalingcooldown where appid = $1"
+	rows, err := sdb.sqldb.Query(query, appId)
+	if err != nil {
+		sdb.logger.Error("can-scale-app-query-record", err, lager.Data{"query": query, "appid": appId})
+		return false, err
+	}
+	defer rows.Close()
+
+	if rows.Next() {
+		var expireAt int64
+		if err = rows.Scan(&expireAt); err != nil {
+			sdb.logger.Error("can-scale-app-scan", err, lager.Data{"query": query, "appid": appId})
+			return false, err
+		}
+		if expireAt < time.Now().UnixNano() {
+			return true, nil
+		} else {
+			return false, nil
+		}
+	}
+
+	return true, nil
+}
+
+func (sdb *ScalingEngineSQLDB) UpdateScalingCooldownExpireTime(appId string, expireAt int64) error {
+	_, err := sdb.sqldb.Exec("DELETE FROM scalingcooldown where appid = $1", appId)
+	if err != nil {
+		sdb.logger.Error("update-scaling-cooldown-time-delete", err, lager.Data{"appid": appId})
+		return err
+	}
+
+	_, err = sdb.sqldb.Exec("INSERT INTO scalingcooldown(appid, expireat) values($1, $2)", appId, expireAt)
+	if err != nil {
+		sdb.logger.Error("update-scaling-cooldown-time-insert", err, lager.Data{"appid": appId, "expireAt": expireAt})
+		return err
+	}
+	return nil
+}
+
+func (sdb *ScalingEngineSQLDB) GetActiveSchedule(appId string) (*models.ActiveSchedule, error) {
+	query := "SELECT scheduleid, instancemincount, instancemaxcount, initialmininstancecount" +
+		" FROM activeschedule WHERE appid = $1"
+
+	var scheduleId string
+	var instanceMin, instanceMax, instanceMinInitial int
+
+	err := sdb.sqldb.QueryRow(query, appId).Scan(&scheduleId, &instanceMin, &instanceMax, &instanceMinInitial)
+	if err == sql.ErrNoRows {
+		return nil, nil
+	}
+	if err != nil {
+		sdb.logger.Error("failed-get-active-schedule-query-row-scan", err, lager.Data{"query": query, "appid": appId})
+		return nil, err
+	}
+
+	return &models.ActiveSchedule{
+		ScheduleId:         scheduleId,
+		InstanceMin:        instanceMin,
+		InstanceMax:        instanceMax,
+		InstanceMinInitial: instanceMinInitial,
+	}, nil
+}
+
+func (sdb *ScalingEngineSQLDB) RemoveActiveSchedule(appId string) error {
+	query := "DELETE FROM activeschedule WHERE appid = $1"
+	_, err := sdb.sqldb.Exec(query, appId)
+	if err != nil {
+		sdb.logger.Error("failed-remove-active-scheudle", err, lager.Data{"appid": appId})
+	}
+	return err
+}
+
+func (sdb *ScalingEngineSQLDB) SetActiveSchedule(appId string, schedule *models.ActiveSchedule) error {
+	err := sdb.RemoveActiveSchedule(appId)
+	if err != nil {
+		sdb.logger.Error("failed-set-active-scheudle-remove", err, lager.Data{"appid": appId})
+		return err
+	}
+
+	query := "INSERT INTO activeschedule(appid, scheduleid, instancemincount, instancemaxcount, initialmininstancecount) " +
+		" VALUES ($1, $2, $3, $4, $5)"
+	_, err = sdb.sqldb.Exec(query, appId, schedule.ScheduleId, schedule.InstanceMin, schedule.InstanceMax, schedule.InstanceMinInitial)
+
+	if err != nil {
+		sdb.logger.Error("failed-set-active-scheudle-insert", err, lager.Data{"appid": appId, "schedule": schedule})
+	}
+	return err
+}
