+++ src/autoscaler/scalingengine/config/config_test.go
@@ -71,8 +71,7 @@ logging:
   level: DeBug
 db:
   policy_db_url: test-policy-db-url
-  history_db_url: test-history-db-url
-  schedule_db_url: test-schedule-db-url
+  scalingengine_db_url: test-scalingengine-db-url
 `)
 			})
 
@@ -91,9 +90,7 @@ db:
 				Expect(conf.Logging.Level).To(Equal("debug"))
 
 				Expect(conf.Db.PolicyDbUrl).To(Equal("test-policy-db-url"))
-				Expect(conf.Db.HistoryDbUrl).To(Equal("test-history-db-url"))
-				Expect(conf.Db.ScheduleDbUrl).To(Equal("test-schedule-db-url"))
-
+				Expect(conf.Db.ScalingEngineDbUrl).To(Equal("test-scalingengine-db-url"))
 			})
 		})
 
@@ -104,8 +101,7 @@ cf:
   api: https://api.example.com
 db:
   policy_db_url: test-policy-db-url
-  history_db_url: test-history-db-url
-  schedule_db_url: test-schedule-db-url
+  scalingengine_db_url: test-scalingengine-db-url
 `)
 			})
 
@@ -127,8 +123,7 @@ db:
 			conf.Cf.GrantType = cf.GrantTypePassword
 			conf.Cf.Username = "admin"
 			conf.Db.PolicyDbUrl = "test-policy-db-url"
-			conf.Db.HistoryDbUrl = "test-history-db-url"
-			conf.Db.ScheduleDbUrl = "test-schedule-db-url"
+			conf.Db.ScalingEngineDbUrl = "test-scalingengine-db-url"
 		})
 
 		JustBeforeEach(func() {
@@ -161,23 +156,13 @@ db:
 			})
 		})
 
-		Context("when history db url is not set", func() {
+		Context("when scalingengine db url is not set", func() {
 			BeforeEach(func() {
-				conf.Db.HistoryDbUrl = ""
+				conf.Db.ScalingEngineDbUrl = ""
 			})
 
 			It("should error", func() {
-				Expect(err).To(MatchError(MatchRegexp("Configuration error: History DB url is empty")))
-			})
-		})
-
-		Context("when schedule db url is not set", func() {
-			BeforeEach(func() {
-				conf.Db.ScheduleDbUrl = ""
-			})
-
-			It("should error", func() {
-				Expect(err).To(MatchError(MatchRegexp("Configuration error: Schedule DB url is empty")))
+				Expect(err).To(MatchError(MatchRegexp("Configuration error: ScalingEngine DB url is empty")))
 			})
 		})
