+++ src/autoscaler/scalingengine/exampleconfig/example.yaml
@@ -9,4 +9,4 @@ logging:
   level: "debug"
 db:
   policy_db_url: "postgres://postgres:postgres@localhost/autoscaler?sslmode=disable"
-  history_db_url: "postgres://postgres:postgres@localhost/autoscaler?sslmode=disable"
+  scalingengine_db_url: "postgres://postgres:postgres@localhost/autoscaler?sslmode=disable"
